Refactor to use modules - more but smaller files
