+++ .eslintrc.js
@@ -2,7 +2,8 @@
 module.exports = {
     "env": {
         "browser": true,
-        "es2021": true
+        "es2021": true,
+        "jquery": true
     },
     "extends": "eslint:recommended",
     "parserOptions": {
