+++ index.html
@@ -14,11 +14,9 @@
         <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.8/dist/umd/popper.min.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+        <script src="https://cdn.jsdelivr.net/npm/chart.js" type="module"></script>
 
-        <script src="src/js/data/achievementInfos.js"></script>
-        <script src="src/js/data/schemas.js"></script>
-        <script src="src/js/script.js"></script>
+        <script src="src/js/loader.js" type="module"></script>
 
         <title>ESO Shared Achievements Viewer</title>
         <link rel="icon" type="image/x-icon" href="favicon.ico">
