+++ src/js/data/viewTableConfig.js
@@ -1,0 +1,19 @@
+"use strict";
+
+export const viewTableConfig = {
+    "dlcDungeons": {
+        "parentElementId": "dlcDungeonsTableBody",
+        "numDataColumnsInMainView": 7,
+        "databaseEntriesType": "dungeon"
+    },
+    "baseDungeons": {
+        "parentElementId": "baseDungeonsTableBody",
+        "numDataColumnsInMainView": 5,
+        "databaseEntriesType": "baseDungeon"
+    },
+    "trials": {
+        "parentElementId": "trialsTableBody",
+        "numDataColumnsInMainView": 6,
+        "databaseEntriesType": "trial"
+    },
+}
