+++ src/js/model/player.js
@@ -1,0 +1,10 @@
+"use strict";
+
+export class Player {
+    constructor() {
+        this.name = "";
+        this.originalInput = "";
+        this.binaryCode = "";
+        this.schemaVersion = -1;
+    }
+}
